scripts/requirements.txt
# Python dependencies for data import scripts
# Install with: pip install -r requirements.txt

supabase>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

